refactor project structure + curiosity base
