+++ locked-requirements.txt
@@ -1,0 +1,218 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile requirements.txt --output-file locked-requirements.txt
+aiosignal==1.3.1
+    # via ray
+attrs==24.2.0
+    # via
+    #   jsonschema
+    #   referencing
+certifi==2024.8.30
+    # via requests
+charset-normalizer==3.3.2
+    # via requests
+click==8.1.7
+    # via
+    #   ray
+    #   typer
+cloudpickle==3.0.0
+    # via gymnasium
+contourpy==1.3.0
+    # via matplotlib
+cycler==0.12.1
+    # via matplotlib
+dm-tree==0.1.8
+    # via ray
+dsse==1.1.9
+    # via -r requirements.txt
+farama-notifications==0.0.4
+    # via gymnasium
+filelock==3.16.1
+    # via
+    #   ray
+    #   torch
+    #   triton
+fonttools==4.54.1
+    # via matplotlib
+frozenlist==1.4.1
+    # via
+    #   aiosignal
+    #   ray
+fsspec==2024.9.0
+    # via
+    #   ray
+    #   torch
+gymnasium==0.28.1
+    # via
+    #   dsse
+    #   pettingzoo
+    #   ray
+idna==3.10
+    # via requests
+imageio==2.35.1
+    # via scikit-image
+jax-jumpy==1.0.0
+    # via gymnasium
+jinja2==3.1.4
+    # via torch
+jsonschema==4.23.0
+    # via ray
+jsonschema-specifications==2023.12.1
+    # via jsonschema
+kiwisolver==1.4.7
+    # via matplotlib
+lazy-loader==0.4
+    # via scikit-image
+llvmlite==0.43.0
+    # via numba
+lz4==4.3.3
+    # via ray
+markdown-it-py==3.0.0
+    # via rich
+markupsafe==2.1.5
+    # via jinja2
+matplotlib==3.8.4
+    # via dsse
+mdurl==0.1.2
+    # via markdown-it-py
+mpmath==1.3.0
+    # via sympy
+msgpack==1.1.0
+    # via ray
+networkx==3.3
+    # via
+    #   scikit-image
+    #   torch
+numba==0.60.0
+    # via dsse
+numpy==2.0.2
+    # via
+    #   contourpy
+    #   dsse
+    #   gymnasium
+    #   imageio
+    #   jax-jumpy
+    #   matplotlib
+    #   numba
+    #   pandas
+    #   pettingzoo
+    #   pyarrow
+    #   scikit-image
+    #   scipy
+    #   tensorboardx
+    #   tifffile
+nvidia-cublas-cu12==12.1.3.1
+    # via
+    #   nvidia-cudnn-cu12
+    #   nvidia-cusolver-cu12
+    #   torch
+nvidia-cuda-cupti-cu12==12.1.105
+    # via torch
+nvidia-cuda-nvrtc-cu12==12.1.105
+    # via torch
+nvidia-cuda-runtime-cu12==12.1.105
+    # via torch
+nvidia-cudnn-cu12==9.1.0.70
+    # via torch
+nvidia-cufft-cu12==11.0.2.54
+    # via torch
+nvidia-curand-cu12==10.3.2.106
+    # via torch
+nvidia-cusolver-cu12==11.4.5.107
+    # via torch
+nvidia-cusparse-cu12==12.1.0.106
+    # via
+    #   nvidia-cusolver-cu12
+    #   torch
+nvidia-nccl-cu12==2.20.5
+    # via torch
+nvidia-nvjitlink-cu12==12.6.68
+    # via
+    #   nvidia-cusolver-cu12
+    #   nvidia-cusparse-cu12
+nvidia-nvtx-cu12==12.1.105
+    # via torch
+packaging==24.1
+    # via
+    #   lazy-loader
+    #   matplotlib
+    #   ray
+    #   scikit-image
+    #   tensorboardx
+pandas==2.2.3
+    # via
+    #   -r requirements.txt
+    #   ray
+pettingzoo==1.24.3
+    # via dsse
+pillow==10.4.0
+    # via
+    #   imageio
+    #   matplotlib
+    #   scikit-image
+protobuf==5.28.2
+    # via
+    #   ray
+    #   tensorboardx
+pyarrow==17.0.0
+    # via ray
+pygame==2.6.1
+    # via dsse
+pygments==2.18.0
+    # via rich
+pyparsing==3.1.4
+    # via matplotlib
+python-dateutil==2.9.0.post0
+    # via
+    #   matplotlib
+    #   pandas
+pytz==2024.2
+    # via pandas
+pyyaml==6.0.2
+    # via ray
+ray==2.37.0
+    # via -r requirements.txt
+referencing==0.35.1
+    # via
+    #   jsonschema
+    #   jsonschema-specifications
+requests==2.32.3
+    # via ray
+rich==13.8.1
+    # via
+    #   ray
+    #   typer
+rpds-py==0.20.0
+    # via
+    #   jsonschema
+    #   referencing
+scikit-image==0.24.0
+    # via ray
+scipy==1.14.1
+    # via
+    #   ray
+    #   scikit-image
+shellingham==1.5.4
+    # via typer
+six==1.16.0
+    # via python-dateutil
+sympy==1.13.3
+    # via torch
+tensorboardx==2.6.2.2
+    # via ray
+tifffile==2024.9.20
+    # via scikit-image
+torch==2.4.1
+    # via -r requirements.txt
+triton==3.0.0
+    # via torch
+typer==0.12.5
+    # via ray
+typing-extensions==4.12.2
+    # via
+    #   gymnasium
+    #   torch
+    #   typer
+tzdata==2024.2
+    # via pandas
+urllib3==2.2.3
+    # via requests
