+++ pyproject.toml
@@ -1,0 +1,92 @@
+[project]
+name = "drone-swarm-search-algorithms"
+version = "0.1.0"
+description = "Algorithms to solve the DSSE environment, focusing on optimizing drone swarm search and navigation for critical applications."
+authors = [
+    { name="Ricardo Ribeiro Rodrigues", email="ricardorr7@al.insper.edu.br" },
+    { name="Renato Laffranchi Falcão", email="renatolf1@al.insper.edu.br" },
+    { name="Pedro Henrique Britto Aragão Andrade", email="pedroa3@al.insper.edu.br" },
+    { name="Jorás Oliveira", email="jorascco@al.insper.edu.br" },
+    { name="Fabricio Barth", email="fabriciojb@insper.edu.br" },
+]
+readme = "README.md"
+requires-python = ">=3.11"
+dependencies = [
+    "aiosignal==1.3.1",
+    "attrs==24.2.0",
+    "certifi==2024.8.30",
+    "charset-normalizer==3.3.2",
+    "click==8.1.7",
+    "cloudpickle==3.0.0",
+    "contourpy==1.3.0",
+    "cycler==0.12.1",
+    "dm-tree==0.1.8",
+    "dsse==1.1.9",
+    "farama-notifications==0.0.4",
+    "filelock==3.16.1",
+    "fonttools==4.54.1",
+    "frozenlist==1.4.1",
+    "fsspec==2024.9.0",
+    "gymnasium==0.28.1",
+    "idna==3.10",
+    "imageio==2.35.1",
+    "jax-jumpy==1.0.0",
+    "jinja2==3.1.4",
+    "jsonschema-specifications==2023.12.1",
+    "jsonschema==4.23.0",
+    "kiwisolver==1.4.7",
+    "lazy-loader==0.4",
+    "llvmlite==0.43.0",
+    "lz4==4.3.3",
+    "markdown-it-py==3.0.0",
+    "markupsafe==2.1.5",
+    "matplotlib==3.8.4",
+    "mdurl==0.1.2",
+    "mpmath==1.3.0",
+    "msgpack==1.1.0",
+    "networkx==3.3",
+    "numba==0.60.0",
+    "numpy==2.0.2",
+    "nvidia-cublas-cu12==12.1.3.1",
+    "nvidia-cuda-cupti-cu12==12.1.105",
+    "nvidia-cuda-nvrtc-cu12==12.1.105",
+    "nvidia-cuda-runtime-cu12==12.1.105",
+    "nvidia-cudnn-cu12==9.1.0.70",
+    "nvidia-cufft-cu12==11.0.2.54",
+    "nvidia-curand-cu12==10.3.2.106",
+    "nvidia-cusolver-cu12==11.4.5.107",
+    "nvidia-cusparse-cu12==12.1.0.106",
+    "nvidia-nccl-cu12==2.20.5",
+    "nvidia-nvjitlink-cu12==12.6.68",
+    "nvidia-nvtx-cu12==12.1.105",
+    "packaging==24.1",
+    "pandas==2.2.3",
+    "pettingzoo==1.24.3",
+    "pillow==10.4.0",
+    "protobuf==5.28.2",
+    "pyarrow==17.0.0",
+    "pygame==2.6.1",
+    "pygments==2.18.0",
+    "pyparsing==3.1.4",
+    "python-dateutil==2.9.0.post0",
+    "pytz==2024.2",
+    "pyyaml==6.0.2",
+    "ray==2.38.0",
+    "referencing==0.35.1",
+    "requests==2.32.3",
+    "rich==13.8.1",
+    "rpds-py==0.20.0",
+    "scikit-image==0.24.0",
+    "scipy==1.14.1",
+    "shellingham==1.5.4",
+    "six==1.16.0",
+    "sympy==1.13.3",
+    "tensorboardx==2.6.2.2",
+    "tifffile==2024.9.20",
+    "torch==2.4.1",
+    "triton==3.0.0",
+    "typer==0.12.5",
+    "typing-extensions==4.12.2",
+    "tzdata==2024.2",
+    "urllib3==2.2.3",
+]
