+++ src/models/__init__.py
@@ -1,0 +1,7 @@
+from .ppo_cnn import PpoCnnModel
+from .ppo_cnn_lstm import PpoCnnLstmModel
+
+__all__ = [
+    "PpoCnnModel",
+    "PpoCnnLstmModel",
+]
