+++ src/search/train/train_ppo_multi.py
@@ -1,11 +1,10 @@
 import os
 import pathlib
 from DSSE import DroneSwarmSearch
-from DSSE.environment.wrappers import AllPositionsWrapper, RetainDronePosWrapper
+from DSSE.environment.wrappers import AllPositionsWrapper
 import ray
-from ray import air
 from ray import tune
-from ray.rllib.algorithms.dqn.dqn import DQNConfig
+from ray.rllib.algorithms.ppo import PPOConfig
 from ray.rllib.env.wrappers.pettingzoo_env import ParallelPettingZooEnv
 from ray.rllib.models import ModelCatalog
 from ray.rllib.models.torch.torch_modelv2 import TorchModelV2
@@ -47,17 +46,11 @@ class MLPModel(TorchModelV2, nn.Module):
             nn.ReLU(),
             # nn.MaxPool2d(kernel_size=2),
             nn.Conv2d(
-                in_channels=16,
-                out_channels=32,
-                kernel_size=(4, 4),
-                stride=(1, 1)
+                in_channels=16, out_channels=32, kernel_size=(4, 4), stride=(1, 1)
             ),
             nn.ReLU(),
             nn.Conv2d(
-                in_channels=32,
-                out_channels=64,
-                kernel_size=(3, 3),
-                stride=(1, 1)
+                in_channels=32, out_channels=64, kernel_size=(3, 3), stride=(1, 1)
             ),
             # nn.MaxPool2d(kernel_size=2),
             nn.Flatten(),
@@ -73,27 +66,32 @@ class MLPModel(TorchModelV2, nn.Module):
             nn.ReLU(),
             nn.LayerNorm(512),
         )
-        
+
         self.unifier = nn.Sequential(
             nn.Linear(512 * 2, 512),
             nn.ReLU(),
             nn.LayerNorm(512),
         )
         self.policy_fn = nn.Linear(512, num_outputs)
+        self.value_fn = nn.Linear(512, 1)
 
     def forward(self, input_dict, state, seq_lens):
         input_ = input_dict["obs"]
         # Convert dims
         input_cnn = input_[1].unsqueeze(1)
         model_out = self.conv1(input_cnn)
-        
+
         scalar_input = input_[0].float()
         scalar_out = self.fc_scalar(scalar_input)
 
         value_input = torch.cat((model_out, scalar_out), -1)
         value_input = self.unifier(value_input)
-        
+
+        self._value_out = self.value_fn(value_input)
         return self.policy_fn(value_input), state
+
+    def value_function(self):
+        return self._value_out.flatten()
 
 
 def env_creator(args):
@@ -104,7 +102,6 @@ def env_creator(args):
         person_initial_position=(10, 10),
     )
     env = AllPositionsWrapper(env)
-    env = RetainDronePosWrapper(env, [(0, 0), (0, 19), (19, 0), (19, 19)])
     return env
 
 
@@ -116,43 +113,41 @@ if __name__ == "__main__":
     register_env(env_name, lambda config: ParallelPettingZooEnv(env_creator(config)))
     ModelCatalog.register_custom_model("MLPModel", MLPModel)
 
-    config = DQNConfig()
-    config = config.environment(env=env_name)
-    config = config.rollouts(num_rollout_workers=6, rollout_fragment_length="auto", num_envs_per_worker=2)
-    config = config.training(
-        train_batch_size=512,
-        grad_clip=None,
-        target_network_update_freq=1,
-        tau=0.005,
-        gamma=0.99999,
-        n_step=1,
-        double_q=True,
-        dueling=False,
-        model={"custom_model": "MLPModel", "_disable_preprocessor_api": True},
-        v_min=-800,
-        v_max=800,
+    config = (
+        PPOConfig()
+        .environment(env=env_name)
+        .rollouts(num_rollout_workers=4, rollout_fragment_length=128)
+        .training(
+            train_batch_size=512,
+            lr=4e-5,
+            gamma=0.99999,
+            lambda_=0.9,
+            use_gae=True,
+            clip_param=0.4,
+            grad_clip=None,
+            entropy_coeff=0.1,
+            vf_loss_coeff=0.25,
+            vf_clip_param=420,
+            sgd_minibatch_size=64,
+            num_sgd_iter=10,
+            model={
+                "custom_model": "MLPModel",
+                "_disable_preprocessor_api": True,
+            },
+        )
+        .experimental(_disable_preprocessor_api=True)
+        .debugging(log_level="ERROR")
+        .framework(framework="torch")
+        .resources(num_gpus=int(os.environ.get("RLLIB_NUM_GPUS", "1")))
     )
-    config = config.exploration(
-        exploration_config={
-            "type": "EpsilonGreedy",
-            "initial_epsilon": 1.0,
-            "final_epsilon": 0.05,
-            "epsilon_timesteps": 350_000,
-        }
-    )
-    config = config.debugging(log_level="ERROR")
-    config = config.framework(framework="torch")
-    config = config.resources(num_gpus=int(os.environ.get("RLLIB_NUM_GPUS", "1")))
-    config = config.experimental(_disable_preprocessor_api=True)
 
     curr_path = pathlib.Path().resolve()
-    run_config = air.RunConfig(
-        stop={"timesteps_total": 10_000_000 if not os.environ.get("CI") else 50000},
+    tune.run(
+        "PPO",
+        name="PPO",
+        stop={"timesteps_total": 5000000 if not os.environ.get("CI") else 50000},
+        checkpoint_freq=10,
+        # local_dir="ray_results/" + env_name,
         storage_path=f"{curr_path}/ray_res/" + env_name,
-        checkpoint_config=air.CheckpointConfig(checkpoint_frequency=10),
+        config=config.to_dict(),
     )
-    tune.Tuner(
-        "DQN",
-        run_config=run_config,
-        param_space=config.to_dict()
-    ).fit()
